# perf: Track indentation during traversal