--- a/markdown/writer_indent.go
+++ b/markdown/writer_indent.go
@@ -1,18 +1,14 @@
 package markdown
 
 import (
-	"bytes"
 	"io"
-
-	"github.com/yuin/goldmark/ast"
 )
 
 // lineIndentWriter wraps io.Writer and adds given indent everytime new line is created .
 type lineIndentWriter struct {
 	io.Writer
 
-	indent                []byte
-	whitespace            []byte
+	id                    indentation
 	firstWriteExtraIndent []byte
 
 	previousCharWasNewLine bool
@@ -22,33 +18,12 @@ func wrapWithLineIndentWriter(w io.Writer) *lineIndentWriter {
 	return &lineIndentWriter{Writer: w, previousCharWasNewLine: true}
 }
 
-func (l *lineIndentWriter) UpdateIndent(node ast.Node, entering bool) {
-	// Recalculate indent.
-	l.indent = l.indent[:0]
+func (l *lineIndentWriter) PushIndent(indent []byte) {
+	l.id.Push(indent)
+}
 
-	p := node
-	if !entering {
-		p = p.Parent()
-	}
-
-	for ; p != nil; p = p.Parent() {
-		if p.Kind() == ast.KindBlockquote {
-			// Prepend, as we go from down.
-			l.indent = append(append([]byte{}, blockquoteChars...), l.indent...)
-			continue
-		}
-
-		if listItem, ok := p.(*ast.ListItem); ok {
-			// Prepend, as we go from down, but don't count first item.
-			l.indent = append(bytes.Repeat(spaceChar, len(listItemMarkerChars(listItem))), l.indent...)
-			continue
-		}
-	}
-
-	// Split whitespace indent from chars.
-	cut := bytes.TrimRight(l.indent, noAllocString(spaceChar))
-	l.whitespace = l.indent[len(cut):]
-	l.indent = cut
+func (l *lineIndentWriter) PopIndent() {
+	l.id.Pop()
 }
 
 func (l *lineIndentWriter) AddIndentOnFirstWrite(add []byte) {
@@ -71,7 +46,7 @@ func (l *lineIndentWriter) Write(b []byte) (n int, _ error) {
 	writtenFromB := 0
 	for i, c := range b {
 		if l.previousCharWasNewLine {
-			ns, err := l.Writer.Write(l.indent)
+			ns, err := l.Writer.Write(l.id.Indent())
 			n += ns
 			if err != nil {
 				return n, err
@@ -99,11 +74,14 @@ func (l *lineIndentWriter) Write(b []byte) (n int, _ error) {
 		}
 
 		// Not a newline, make a space if indent was created.
-		if l.previousCharWasNewLine && len(l.whitespace) > 0 {
-			ns, err := l.Writer.Write(l.whitespace)
-			n += ns
-			if err != nil {
-				return n, err
+		if l.previousCharWasNewLine {
+			ws := l.id.Whitespace()
+			if len(ws) > 0 {
+				ns, err := l.Writer.Write(ws)
+				n += ns
+				if err != nil {
+					return n, err
+				}
 			}
 		}
 		l.previousCharWasNewLine = false